Unify sorted/unsorted activity model
activity db model contains method to fetch
- sorted activities
- unsorted activities
